Untangle node->colour from anipanel and make the simulation host types responsible
AnimatableHost: implement coloured in the abstract and get rid of the
generic 'type'

simhost: Implemented Coloured, using an enum class and allow each instance
to have a notion of this enum type associated with it.

simapp: the anipanel Node enum is now in simhost.

anipanel: doesn't need to know anything other than a simulation uses
AnimatableHost's, which implement Coloured.